requirements.txt
pytorch_lightning==1.6.0
numpy~=1.22.3
torch~=1.11.0
torchvision~=0.12.0
pyyaml  # 6.0
transformers~=4.18.0
pandas~=1.4.2
pillow~=9.1.0
tqdm~=4.64.0
torchmetrics~=0.8.0